// src/types/invoiceData.d.ts
export type TaxNumberType = {
  taxpayerId: string; // Az adóalany adó törzsszáma. Csoportos adóalany esetén csoportazonosító szám
  vatCode?: string; // Áfakód az adóalanyiság típusanak jelzésére. Egy számjegy
  countyCode?: string; // Megyekód, két számjegy
};
// ...
export type AddressType = {
  simpleAddress?: SimpleAddressType;
  detailedAddress?: DetailedAddressType;
};
// ...
export type InvoiceReferenceType = {
  originalInvoiceNumber: string; // Az eredeti számla sorszáma, melyre a módosítás vonatkozik
  modifyWithoutMaster: boolean; // Annak jelzése, hogy a módosítás olyan alapszámlára hivatkozik, amelyről nem történt és nem is fog történni adatszolgáltatás
  modificationIndex: number; // A számlára vonatkozó módosító okirat egyedi sorszáma
};

export type SupplierInfoType = {
  supplierTaxNumber: TaxNumberType; // Belföldi adószám, amely alatt a számlán szereplő termékértékesítés vagy szolgáltatásnyújtás történt. Lehet csoportazonosító szám is.
  groupMemberTaxNumber?: TaxNumberType; // Csoporttag adószáma, ha a termékértékesítés vagy szolgáltatásnyújtás csoportazonosító szám alatt történt
  communityVatNumber?: string; // Közösségi adószám
  supplierName: string; // Az eladó (szállító) neve
  supplierAddress: AddressType; // Az eladó (szállító) címe
  supplierBankAccountNumber?: string; // Az eladó (szállító) bankszámlaszáma;
  individualExemption?: boolean; // Értéke true, ha a számlakibocsátó (eladó) alanyi áfamentes
  exciseLicenceNum?: string; // Az eladó adóraktári engedélyének vagy jövedéki engedélyének száma
};

export type CustomerVatStatusType = 'DOMESTIC' | 'OTHER' | 'PRIVATE_PERSON';

export type CustomerInfoType = {
  customerVatStatus: CustomerVatStatusType; // Vevő áfa szerinti státusza
  customerVatData?: CustomerVatDataType; // A vevő áfaalanyisági adatai
  customerName?: string; // A vevő neve
  customerAddress?: AddressType; // A vevő címe
  customerBankAccountNumber?: string; // Vevő bankszámlaszáma
};

export type FiscalRepresentativeInfoType = {
  fiscalRepresentativeTaxNumber: TaxNumberType; // A pénzügyi képviselő adószáma
  fiscalRepresentativeName: string; // A pénzügyi képviselő neve
  fiscalRepresentativeAddress: AddressType; // Pénzügyi képviselő címe
  fiscalRepresentativeBankAccountNumber?: string; // Pénzügyi képviselő által a számlakibocsátó (eladó) számára megnyitott bankszámla bankszámlaszáma
};

export type InvoiceCategoryType = 'NORMAL' | 'SIMPLIFIED' | 'AGGREGATE';

export type PaymentMethodType = 'TRANSFER' | 'CASH' | 'CARD' | 'VOUCHER' | 'OTHER';

export type InvoiceAppearanceType = 'PAPER' | 'ELECTRONIC' | 'EDI' | 'UNKNOWN';

export type InvoiceDetailType = {
  invoiceCategory: InvoiceCategoryType; // A számla típusa, módosító okirat esetén az eredeti számla típusa
  invoiceDeliveryDate: string; // Teljesítés dátuma (ha nem szerepel a számlán, akkor azonos a számla keltével)
  invoiceDeliveryPeriodStart?: string; // Ha a számla egy időszakra vonatkozik, akkor az időszak első napja
  invoiceDeliveryPeriodEnd?: string; // Ha a számla egy időszakra vonatkozik, akkor az időszak utolsó napja
  invoiceAccountingDeliveryDate?: string; // Számviteli teljesítés dátuma. Időszak esetén az időszak utolsó napja
  periodicalSettlement?: boolean; // Időszakos elszámolás jelzése
  smallBusinessIndicator?: boolean; // Kisadózó jelzése
  currencyCode: string; // A számla pénzneme az ISO 4217 szabvány szerint
  exchangeRate: number; // HUF-tól különböző pénznem esetén az alkalmazott árfolyam: egy egység értéke HUF-ban
  utilitySettlementIndicator?: boolean; // Közmű elszámolószámla jelölése
  selfBillingIndicator?: boolean; // Önszámlázás jelölése (önszámlázás esetén true)
  paymentMethod?: PaymentMethodType; // Fizetés módja
  paymentDate?: string; // Fizetési határidő
  cashAccountingIndicator?: boolean; // Pénzforgalmi elszámolás jelölése, ha az szerepel a számlán
  invoiceAppearance: InvoiceAppearanceType; // A számla vagy módosító okirat megjelenési formája
  conventionalInvoiceInfo?: ConventionalInvoiceInfoType; // A számlafeldolgozást segítő, egyezményesen nevesített egyéb adatok
  additionalInvoiceData?: AdditionalDataType[]; // A számlára vonatkozó egyéb adat;
};

export type InvoiceHeadType = {
  supplierInfo: SupplierInfoType; // Számlakibocsátó (eladó) adatai
  customerInfo?: CustomerInfoType; //  Vevő adatai
  fiscalRepresentativeInfo?: FiscalRepresentativeInfoType; //  Pénzügyi képviselő adatai
  invoiceDetail: InvoiceDetailType; // Számla részletező adatok
};
// ...
export type ProductFeeDataType = {
  productFeeCode: ProductCodeType; // Termékdíj kód (Kt vagy Csk)
  productFeeQuantity: number; // A termékdíjjal érintett termék mennyisége;
  productFeeMeasuringUnit: ProductFeeMeasuringUnitType; //A díjtétel egysége (kg vagy darab)
  productFeeRate: number; // A termékdíj díjtétele (HUF/egység)
  productFeeAmount: number; // Termékdíj összege forintban
};

export type LineType = {
  lineNumber: number; // A tétel sorszáma
  lineModificationReference?: LineModificationReferenceType; // Módosító számla esetén a tételsorszintű módosítások jelölése
  referencesToOtherLines?: // Hivatkozások kapcsolódó tételekre, ha ez az Áfa törvény alapján szükséges
  {
    referenceToOtherLine: number[]; //Hivatkozások kapcsolódó tételekre, ha ez az Áfa törvény alapján szükséges
  };

  advanceData?: AdvanceDataType; // Előleghez kapcsolódó adatok
  productCodes?: // Termékkódok
  {
    productCode: ProductCodeType[]; // Termékkód
  };

  lineExpressionIndicator: boolean; // Értéke true, ha a tétel mennyiségi egysége természetes mértékegységben kifejezhető
  lineNatureIndicator?: LineNatureIndicatorType; // Termékértékesítés vagy szolgáltatásnyújtás jelölése
  lineDescription?: string; // A termék vagy szolgáltatás megnevezése;
  quantity?: number; // Mennyiség
  unitOfMeasure?: UnitOfMeasureType; // Mennyiség egység
  unitOfMeasureOwn?: string; // Saját mennyiségi egység
  unitPrice?: number; // Egységár a számla pénznemében. Egyszerűsített számla esetén bruttó, egyéb esetben nettó egységár
  unitPriceHUF?: number; // Egységár forintban. Egyszerűsített számla esetén bruttó, egyéb esetben nettó egységár.
  lineDiscountData?: LineDiscountDataType; // A tételhez tartozó árengedmény adatok;
  lineAmountsNormal?: LineamountsNormalType; // Normál (nem egyszerűsített) számla esetén (beleértve a gyűjtőszámlát) kitöltendő tétel érték adatok.
  lineAmountsSimplified?: LineAmountsSimplified; // Egyszerűsített számla esetén kitöltendő tétel érték adatok
  intermediatedService?: boolean; // Értéke true ha a tétel közvetített szolgáltatás
  aggregateInvoiceLineData?: AggregateInvoiceLineData; // Gyűjtőszámla adatok
  newTransportMean?: NewTransportMeanType; // Új közlekedési eszköz értékesítés;
  depositIndicator?: boolean; // Értéke true, ha a tétel betétdíj jellegű;
  obligatedForProductFee?: boolean; // Értéke true ha a tételt termékdíj fizetési kötelezettség terheli
  GPCExcise?: number; // Földgáz, villamos energia, szén jövedéki adója forintban
  dieselOilPurchase?: DieselOilPurchaseType; // Gázolaj adózottan történő beszerzésének adatai
  netaDeclaration?: boolean; //  Értéke true, ha a Neta tv-ben meghatározott adókötelezettség az adó alanyát terheli
  productFeeClause?: ProductFeeClauseType; // A környezetvédelmi termékdíjról szóló 2011. évi LXXXV. tv. szerinti tételekre vonatkozó záradékok
  lineProductFeeContent?: ProductFeeDataType; // A tétel termékdíj tartalmára vonatkozó adatok
  conventionalLineInfo?: ConventionalInvoiceInfoType; // A számlafeldolgozást segítő, egyezményesen nevesített, egyéb adatok
  additionalLineData?: AdditionalDataType; // A termék/szolgáltatás tételhez kapcsolódó, további adat
};

export type LinesType = {
  mergedItemIndicator: boolean; // Jelöli, ha az adatszolgáltatás méretcsökkentés miatt összevont soradatokat tartalmaz
  line?: LineType[]; // Termék/szolgáltatás tétel
};
// ...
export type SummaryNormalType = {
  summaryByVatRate: SummaryByVatRateType[]; // Összesítés áfa-mérték szerint
  invoiceNetAmount: number; // A számla nettó összege a számla pénznemében;
  invoiceNetAmountHUF: number; // A számla nettó összege forintban;
  invoiceVatAmount: number; // A számla áfa összege a számla pénznemében;
  invoiceVatAmountHUF: number; // A számla áfa összege forintban
};

export type SummaryGrossDataType = {
  invoiceGrossAmount: number; // A számla bruttó összege a számla pénznemében
  invoiceGrossAmountHUF: number; // A számla bruttó összege forintban;
};

export type SummarySimplifiedType = {
  vatRate: VatRateType; // Egyszerűsített számla esetén az adótartalom vagy adómentesség jelölése
  vatContentGrossAmount: number; // Az adott adótartalomhoz tartozó értékesítés vagy szolgáltatásnyújtás bruttó összege a számla pénznemében
  vatContentGrossAmountHUF: number; // Az adott adótartalomhoz tartozó értékesítés vagy szolgáltatásnyújtás bruttó összege forintban
};

export type InvoiceSummaryType = {
  summaryNormal?: SummaryNormalType;
  summarySimplified?: SummarySimplifiedType[];
  summaryGrossData?: SummaryGrossDataType;
};

export type ProductFeeOperationType = 'REFUND' | 'DEPOSIT';

export type ProductFeeSummaryType = {
  productFeeOperation: ProductFeeOperationType; // Annak jelzése, hogy a termékdíj összesítés visszaigénylésre (REFUND) vagy raktárba történő beszállításra (DEPOSIT) vonatkozik
  productFeeData: ProductFeeDataType[]; // Termékdíjadatok
  ProductChargeSum: number; // Termékdíj összesen
  PaymentEvidenceDocumentData?: // A termékdíj bevallását igazoló dokumentum adatai a 2011. évi
  {
    evidenceDocumentNo: string; // Számla sorszáma vagy egyéb okirat azonosító száma
    evidenceDocumentDate: string; // Számla kelte
    obligatedName: string; // Kötelezett neve
    obligatedAddress: AddressType; // Kötelezett címe
    obligatedTaxNumber: TaxNumberType; // A kötelezett adószáma
  };
};

export type InvoiceType = {
  invoiceReference?: InvoiceReferenceType;
  invoiceHead: InvoiceHeadType;
  invoiceLines?: LinesType;
  productFeeSummary?: ProductFeeSummaryType[];
  invoiceSummary: InvoiceSummaryType;
};

export type BatchInvoiceType = {
  batchIndex: number;
  invoice: InvoiceType;
};

export type InvoiceMainType =
  | ({ invoice: InvoiceType } & { batchInvoice?: never })
  | ({ batchInvoice: BatchInvoiceType[] } & { invoice?: never });

type Namespaces = {
  '$xsi:schemaLocation': 'http://schemas.nav.gov.hu/OSA/3.0/data invoiceData.xsd';
  $xmlns: 'http://schemas.nav.gov.hu/OSA/3.0/data';
  '$xmlns:xsi': 'http://www.w3.org/2001/XMLSchema-instance';
  '$xmlns:common': 'http://schemas.nav.gov.hu/NTCA/1.0/common';
  '$xmlns:base': 'http://schemas.nav.gov.hu/OSA/3.0/base';
};

export type InvoiceDataProps = {
  invoiceNumber: string; // Számla vagy módosító okirat sorszáma;
  invoiceIssueDate: string; // Számla vagy módosító okirat kelte
  completenessIndicator: boolean; // Jelöli, ha az adatszolgáltatás maga a számla
  invoiceMain: InvoiceMainType;
} & Namespaces;

export type InvoiceData = {
  InvoiceData: InvoiceDataProps;
};